requirements.txt
Flask
Jinja2
requests
google-generativeai
gunicorn